Add files via upload
SS1 is the output when no file "department.txt." exists yet, so when the code is run for the first time.

SS2 shows the process of creating a department. When done, it displays a message to say if it was successful, and then displays the menu again.

SS3 shows the process of adding a teacher to a department. The program asks for input where it is printed "Enter" (or when it asks if the teacher is full time). When all input has been given, the program displays a message to say that the process was successful. It then shows the menu once again.

SS4 shows the process of saving the data (creating "departments.txt").  The menu is then displayed again.

SS5 shows the inside of the "departments.txt" file, which was created by the program, storing all information about every departments and their teachers and staff.

SS6 shows the output when the program is run once again. The program shows the existing departments from file "departments.txt" (here the Dean ID is -1 because there is no dean). Then the menu is displayed once again.

SS9 shows the process of viewing the departments. When choosing option 5, the program shows all departments in the system with their teachers and staff members (and their informations). This proves that the data was appropriately read from the "department.txt" file.

SS10 shows the process of adding a staff member. It is similar to adding a teacher. The menu is yet again displayed at the end.

SS11 shows the process of setting a dean. It allows to chose a teacher from selected department. Then the menu is displayed.